requirements/base.txt
torch>=2.0
numpy>=1.23
nibabel>=5.1
scikit-learn>=1.3
matplotlib>=3.7
tqdm>=4.66
PyYAML>=6.0
jsonschema>=4.21
psutil>=5.9
